refactor: recreate proxies for feature-management
