+++ npm/ng-packs/apps/dev-app/src/environments/environment.ts
@@ -22,5 +22,9 @@ export const environment = {
       url: 'https://localhost:44305',
       rootNamespace: 'MyCompanyName.MyProjectName',
     },
+    AbpFeatureManagement: {
+      url: 'https://localhost:44305',
+      rootNamespace: 'Volo.Abp',
+    },
   },
 } as Config.Environment;
