+++ npm/ng-packs/packages/feature-management/src/lib/proxy/feature-management/features.service.ts
@@ -1,0 +1,29 @@
+import type { GetFeatureListResultDto, UpdateFeaturesDto } from './models';
+import { RestService } from '@abp/ng.core';
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root',
+})
+export class FeaturesService {
+  apiName = 'AbpFeatureManagement';
+
+  get = (providerName: string, providerKey: string) =>
+    this.restService.request<any, GetFeatureListResultDto>({
+      method: 'GET',
+      url: `/api/feature-management/features`,
+      params: { providerName: providerName, providerKey: providerKey },
+    },
+    { apiName: this.apiName });
+
+  update = (providerName: string, providerKey: string, input: UpdateFeaturesDto) =>
+    this.restService.request<any, void>({
+      method: 'PUT',
+      url: `/api/feature-management/features`,
+      params: { providerName: providerName, providerKey: providerKey },
+      body: input,
+    },
+    { apiName: this.apiName });
+
+  constructor(private restService: RestService) {}
+}
