+++ npm/ng-packs/packages/feature-management/src/lib/proxy/index.ts
@@ -1,1 +1,2 @@
-export * from './volo';
+export * as FeatureManagement from './feature-management';
+export * as Validation from './validation';
