+++ npm/ng-packs/packages/feature-management/src/lib/proxy/validation/index.ts
@@ -1,0 +1,1 @@
+export * as StringValues from './string-values';
