+++ npm/ng-packs/packages/feature-management/src/lib/proxy/validation/string-values/index.ts
