+++ npm/ng-packs/packages/feature-management/src/lib/proxy/validation/string-values/models.ts
@@ -1,0 +1,13 @@
+
+export interface IStringValueType {
+  name: string;
+  item: object;
+  properties: Record<string, object>;
+  validator: IValueValidator;
+}
+
+export interface IValueValidator {
+  name: string;
+  item: object;
+  properties: Record<string, object>;
+}
