--- npm/ng-packs/packages/feature-management/src/lib/proxy/volo/abp/index.ts
@@ -1,2 +1,0 @@
-export * from './feature-management';
-export * from './validation';
