--- npm/ng-packs/packages/feature-management/src/lib/proxy/volo/abp/validation/index.ts
@@ -1,1 +1,0 @@
-export * from './string-values';
