--- npm/ng-packs/packages/feature-management/src/lib/proxy/volo/index.ts
@@ -1,1 +1,0 @@
-export * from './abp';
