+++ npm/ng-packs/packages/feature-management/src/public-api.ts
@@ -1,4 +1,5 @@
 export * from './lib/feature-management.module';
 export * from './lib/components';
 export * from './lib/enums/components';
-export * from './lib/proxy';
+export * from './lib/proxy/feature-management';
+export * from './lib/proxy/validation/string-values';
